requirements.txt
amqp==5.0.5
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chromedriver-autoinstaller==0.2.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses
decorator==4.4.2
defusedxml==0.7.0
Django==3.0.5
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangotoolbox==1.8.0
djongo==1.3.4
dnspython==1.16.0
entrypoints==0.3
eventlet==0.30.2
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kombu==5.0.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
openapi-codec==1.3.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
redis==3.5.3
requests==2.25.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.2.4
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
python-environ==0.4.54
gunicorn==20.1.0